+++ src/screen/dashboard.rs
@@ -44,6 +44,7 @@ pub struct Dashboard {
     command_bar: Option<CommandBar>,
     file_transfers: file_transfer::Manager,
     theme_editor: Option<ThemeEditor>,
+    notifications: notification::Notifications,
 }
 
 #[derive(Debug)]
@@ -88,6 +89,7 @@ impl Dashboard {
             command_bar: None,
             file_transfers: file_transfer::Manager::new(config.file_transfer.clone()),
             theme_editor: None,
+            notifications: notification::Notifications::new(),
         };
 
         let command = dashboard.track();
@@ -1872,7 +1874,11 @@ impl Dashboard {
             .file_transfers
             .receive(request.clone(), config.proxy.as_ref())?;
 
-        notification::file_transfer_request(&config.notifications, request.from.clone(), server);
+        self.notifications.notify(
+            &config.notifications,
+            &client::Notification::FileTransferRequest(request.from.clone()),
+            Some(server),
+        );
 
         let query =
             target::Query::parse(request.from.as_ref(), chantypes, statusmsg, casemapping).ok()?;
@@ -1960,6 +1966,7 @@ impl Dashboard {
             command_bar: None,
             file_transfers: file_transfer::Manager::new(config.file_transfer.clone()),
             theme_editor: None,
+            notifications: notification::Notifications::new(),
         };
 
         let mut tasks = vec![];